refractor: improve data transfer and storage logic

+++ src/main/java/com/project/hearmeout_backend/repository/PostRepository.java
@@ -12,6 +12,6 @@ public interface PostRepository extends JpaRepository<@NonNull Post,@NonNull  Lo
 
     List<Post> findByAuthor_IdAndPostType(Long userId, PostType postType);
 
-    List<Post> findByPostTypeAndAuthorIdNot(PostType postType, Pageable pageable, int userId);
+    List<Post> findByPostTypeAndAuthorIdNot(PostType postType, Pageable pageable, Long userId);
     List<Post> findByPostType(PostType postType, Pageable pageable);
 }
